requirements.txt
aioredis==1.3.1
amqp==5.0.9
arabic-reshaper==2.1.3
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.3.2
Automat==20.2.0
billiard==3.6.4.0
boto3==1.26.7
botocore==1.29.7
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
charset-normalizer==2.0.11
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.29.0
constantly==15.1.0
cryptography==36.0.2
daphne==3.0.2
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2.9
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-cloudinary-storage==0.3.0
django-crispy-forms==1.13.0
django-elasticsearch-dsl==7.2.0
django-filter==21.1
django-geojson==3.2.0
django-otp==1.1.3
django-redis==5.2.0
django-taggit==2.0.0
django-timezone-field==4.2.3
django-tinymce==3.3.0
djangorestframework==3.13.1
djangorestframework-gis==0.17
elasticsearch==7.13.4
elasticsearch-dsl==7.4.0
future==0.18.2
gunicorn==20.1.0
hiredis==2.0.0
html5lib==1.1
humanize==4.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
jmespath==1.0.1
kombu==5.2.3
msgpack==1.0.3
packaging==21.3
Pillow==9.0.1
prometheus-client==0.13.1
prompt-toolkit==3.0.28
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycurl==7.44.1
pyOpenSSL==22.0.0
pyparsing==3.0.7
PyPDF2==1.26.0
python-bidi==0.4.2
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
qrcode==7.3.1
redis==4.1.4
reportlab==3.5.64
requests==2.26.0
s3transfer==0.6.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
taggit-selectize==2.11.0
tornado==6.1
Twisted==22.2.0
txaio==22.2.1
typing_extensions==4.1.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==6.0.0
wrapt==1.13.3
xhtml2pdf==0.2.5
zope.interface==5.4.0
